fix state management when use continues editing unauthorized NW
Imporove naming/oranization of modules
