+++ ui/package.json
@@ -9,8 +9,10 @@
     "env-cmd": "^8.0.2",
     "flux": "^3.1.3",
     "history": "^4.7.2",
+    "install": "^0.12.1",
     "jwt-decode": "^2.2.0",
     "leaflet": "^1.2.0",
+    "npm": "^6.4.0",
     "prop-types": "^15.6.0",
     "query-string": "^6.1.0",
     "ramda": "^0.25.0",
